Add SM Partial Support

+++ src/base/RAConst.py
@@ -14,26 +14,26 @@ class RAConst :
     MSEC_TO_SEC : float = 1.0 / 1000.0
 
     @staticmethod
-    def hrToMin(hours) : return hours * RAConst.HR_TO_MIN
+    def hrToMin(hours): return float(hours * RAConst.HR_TO_MIN)
     @staticmethod
-    def hrToSec(hours) : return hours * RAConst.HR_TO_SEC
+    def hrToSec(hours): return float(hours * RAConst.HR_TO_SEC)
     @staticmethod
-    def hrToMSec(hours) : return hours * RAConst.HR_TO_MSEC
+    def hrToMSec(hours): return float(hours * RAConst.HR_TO_MSEC)
     @staticmethod
-    def minToHr(mins) : return mins * RAConst.MIN_TO_HR
+    def minToHr(mins): return float(mins * RAConst.MIN_TO_HR)
     @staticmethod
-    def minToSec(mins) : return mins * RAConst.MIN_TO_SEC
+    def minToSec(mins): return float(mins * RAConst.MIN_TO_SEC)
     @staticmethod
-    def minToMSec(mins) : return mins * RAConst.MIN_TO_MSEC
+    def minToMSec(mins): return float(mins * RAConst.MIN_TO_MSEC)
     @staticmethod
-    def secToHr(secs) : return secs * RAConst.SEC_TO_HR
+    def secToHr(secs): return float(secs * RAConst.SEC_TO_HR)
     @staticmethod
-    def secToMin(secs) : return secs * RAConst.SEC_TO_MIN
+    def secToMin(secs): return float(secs * RAConst.SEC_TO_MIN)
     @staticmethod
-    def secToMSec(secs) : return secs * RAConst.SEC_TO_MSEC
+    def secToMSec(secs): return float(secs * RAConst.SEC_TO_MSEC)
     @staticmethod
-    def mSecToHr(msecs) : return msecs * RAConst.MSEC_TO_HR
+    def mSecToHr(msecs): return float(msecs * RAConst.MSEC_TO_HR)
     @staticmethod
-    def mSecToMin(msecs) : return msecs * RAConst.MSEC_TO_MIN
+    def mSecToMin(msecs): return float(msecs * RAConst.MSEC_TO_MIN)
     @staticmethod
-    def mSecToSec(msecs) : return msecs * RAConst.MSEC_TO_SEC
+    def mSecToSec(msecs): return float(msecs * RAConst.MSEC_TO_SEC)
